requirements.txt
requests
pytest
structlog
curlify
retrying
pydantic
pyhamcrest
vyper-config
allure-pytest